requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
